Move matcher classes to new package
