+++ validator/src/main/java/com/github/maracas/validator/matchers/LocationMatcher.java
@@ -1,4 +1,4 @@
-package com.github.maracas.validator.accuracy;
+package com.github.maracas.validator.matchers;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -11,6 +11,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import com.github.maracas.brokenUse.BrokenUse;
+import com.github.maracas.validator.accuracy.AccuracyCase;
 import com.github.maracas.validator.accuracy.AccuracyCase.AccuracyType;
 import com.github.maracas.validator.build.CompilerMessage;
